pyproject.toml
[tool.poetry]
name = "django-benkyou"
version = "0.1.0"
description = ""
authors = ["XERSUY <dev.luyoes20@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
django-ninja = "^0.17.0"
PyJWT = "^2.3.0"
virtualenv = "^20.14.0"
SQLAlchemy = "^1.4.34"
python-dotenv = "^0.20.0"
sshtunnel = "^0.4.0"
psycopg2-binary = "^2.9.3"
pandas = "^1.4.2"
simplejson = "^3.17.6"
pendulum = "^2.1.2"
numpy = "^1.22.3"
DateTime = "^4.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
